added logging and dto

+++ OpenAlprDotNetService/Dto/RecognitionResult.cs
@@ -6,8 +6,8 @@ namespace OpenAlprDotNetService.Dto
     public class RecognitionResult
     {
         public string PlateNumber { get; set; }
-        public float Confidence { get; set; }
-        public float TotalProcessingTimeMs { get; set; }
+        public double Confidence { get; set; }
+        public double TotalProcessingTimeMs { get; set; }
         public IEnumerable<PlatePoint> PlatePoints { get; set; }
     }
 }
